restructure file format; put engine in new module. edit readme. add contributing.md
