+++ CONTRIBUTING.md
@@ -1,0 +1,28 @@
+# Contributing Guidelines
+
+Thanks for taking the time to contribute!
+
+The following is a set of guidelines for contributing to the Alienation game. These are just guidelines, not rules, so use your best judgement and feel free to propose changes to this document in a pull request.
+
+## Getting Started
+
+Alienation is built in a homemade framework in [Java 11](https://www.oracle.com/java/technologies/javase-jdk11-downloads.html) and tested with [JUnit 5](https://junit.org/junit5/). If you are new to Java, it is recommended you take some MOOCs over at [Udemy](https://www.udemy.com/courses/search/?q=java), [YouTube](https://www.youtube.com/results?search_query=java+11), or on any preferred platform in the learning style that suits you best.
+
+## Issue
+Ensure the bug was not already reported by searching on GitHub under [issues](https://github.com/Capstone-Mafia/Alienation/issues). If you're unable to find an open issue addressing the bug, open a [new issue](https://github.com/Capstone-Mafia/Alienation/issues/new/choose).
+
+Please pay attention to the following points while opening an issue.
+
+### Write detailed information
+Detailed information is very helpful to understand an issue.
+
+For example:
+* How to reproduce the issue, step-by-step.
+* The expected behavior (or what is wrong).
+* Screenshots of the issue.
+
+
+## Pull Requests
+Pull Requests are always welcome. 
+
+- When you edit the code in your IDE, please run `rebuild module` and `debug`to check for errors in the formatting of your code before you `git commit`.
