+++ README.md
@@ -2,4 +2,23 @@
 Text-Based console game using JAVA language.  Character explores rooms on a spaceship trying to survive while avoiding attacks from aliens.
 
 # Authors
-Dhruti Kosta, Terrell Douglas, Brad Smialek
+<b>Original Authors:</b> Dhruti Kosta, Terrell Douglas, Brad Smialek
+
+<b>Contributing Authors:</b> Bruce West, Gurinder Batth, Daeun Lok
+
+## Tech/framework used
+<b>Built with</b>
+- [Java JDK 11](https://www.oracle.com/java/technologies/javase-jdk11-downloads.html)
+- [JUnit 5](https://junit.org/junit5/)
+
+## Installation
+This game is free and open. Fork our repository and hack away. Feel free to take it, run with it, and build your own text adventure games off of our humble little platform.
+
+## Tests
+Install JAva JDK 11 and JUnit 5 (links above). Recommend the intelliJ IDEA IDE for your development environment, as that is what the project was built on, and we know it works there.
+
+## How to use?
+A JAR file will be uploaded in time. One option you may pursue if you seek to play our daringly adventurous game is to clone our repo, and either run it from inside your IDE, or create your own executable JAR file.
+
+## Contribute
+Let people know how they can contribute into your project. A [contributing guideline](https://github.com/zulip/zulip-electron/blob/master/CONTRIBUTING.md) will be a big plus.
