+++ src/com/alienation/coregamefiles/Actions.java
@@ -1,4 +1,4 @@
-package com.alienation.game;
+package com.alienation.coregamefiles;
 
 /**
  * Action Enums
