+++ src/com/alienation/coregamefiles/CapsuleRoom.java
@@ -1,6 +1,6 @@
-package com.alienation.game;
+package com.alienation.coregamefiles;
 
-import static com.alienation.game.TextColors.*;
+import static com.alienation.coregamefiles.TextColors.*;
 
 /**
  * Capsule Room - This is the room where the character wakes up
