+++ src/com/alienation/coregamefiles/RoomsMap.java
@@ -1,4 +1,4 @@
-package com.alienation.game;
+package com.alienation.coregamefiles;
 
 /**
  * grabs map depending on which room character is in
