+++ tests/com/alienation/coregamefiles/InventoryTest.java
@@ -1,14 +1,12 @@
-package com.alienation.game;
+package com.alienation.coregamefiles;
 
 import org.junit.After;
 import org.junit.Before;
 
-import static org.junit.Assert.*;
-
 /**
  * Created by bradsmialek on Thu - 7/16/20 @ 12:45 PM
  */
-public class OxygenTest {
+public class InventoryTest {
 
     @Before
     public void setUp() throws Exception {
